# apps/docs/src/data/openapi/nest-auth.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "Register a new user. Response format depends on accessTokenType configuration:\n- Header mode (default): Returns tokens in response body\n- Cookie mode: Sets tokens in HTTP-only cookies and returns success message",
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthSignupRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Header mode: Returns message + tokens in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Signup",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate user. Response format depends on accessTokenType configuration:\n- Header mode (default): Returns tokens in response body\n- Cookie mode: Sets tokens in HTTP-only cookies and returns success message",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthLoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Header mode: Returns message + tokens in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/passwordless/send": {
      "post": {
        "operationId": "AuthController_passwordlessSend",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthPasswordlessSendRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Passwordless — send login code (email or SMS)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "description": "Refresh access token. Response format depends on accessTokenType configuration:\n- Header mode (default): Returns new tokens in response body\n- Cookie mode: Sets new tokens in HTTP-only cookies and returns success message",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthRefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Header mode: Returns message + tokens in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Refresh Token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/mfa/challenge": {
      "post": {
        "operationId": "MfaController_challenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthSendMfaCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthMfaCodeSentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send MFA code for setup/verification",
        "tags": [
          "Mfa"
        ]
      }
    },
    "/auth/mfa/verify": {
      "post": {
        "description": "Verify multi-factor authentication. Response format depends on accessTokenType configuration:\n- Header mode (default): Returns tokens in response body\n- Cookie mode: Sets tokens in HTTP-only cookies and returns success message",
        "operationId": "AuthController_verify2fa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthVerify2faRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Header mode: Returns message + tokens in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verify2faWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify MFA",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/switch-tenant": {
      "post": {
        "operationId": "AuthController_switchTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthSwitchTenantRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Switch Active Tenant",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthChangePasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Change Password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthForgotPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthPasswordResetLinkSentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Forgot password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-forgot-password-otp": {
      "post": {
        "operationId": "AuthController_verifyForgotPasswordOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthVerifyForgotPasswordOtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOtpResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify Forgot Password OTP and get reset token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthResetPasswordWithTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthPasswordResetResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/send-email-verification": {
      "post": {
        "operationId": "AuthController_sendEmailVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthSendEmailVerificationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthEmailVerificationSentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send email verification",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthVerifyEmailRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthEmailVerifiedResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify Email",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "@ackplus/nest-auth API",
    "description": "OpenAPI specification generated from the Nest Auth module",
    "version": "2.0.0-beta.16",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "NestAuthSignupRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address (required if phone not provided)",
            "example": "user@example.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone number (required if email not provided)",
            "example": "+1234567890"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecurePass123!",
            "minLength": 8
          },
          "tenantId": {
            "type": "string",
            "description": "Tenant ID for multi-tenant applications",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "guard": {
            "type": "string",
            "description": "Guard context (e.g. admin, web, vendor) for isolation. Deprecated: use client",
            "example": "admin",
            "deprecated": true
          }
        },
        "required": [
          "password"
        ]
      },
...
      "NestAuthPasswordlessSendRequestDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email or phone (per `channel`)",
            "example": "user@example.com"
          },
          "channel": {
            "type": "string",
            "enum": [
              "email",
              "sms"
            ]
          },
          "tenantId": {
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "channel"
        ]
      },
...
    }
  }
}